Switch from Dvorak to ADORE and redo the layout a bit
What used to be the bottom row, mostly moved to the top, because my hands keep
gravitating to the bottom. Switching to ADORE fits the layout of the Atreus just
right, as it leaves the key right beside the big thumb keys reusable, and I can
place backspace and space on those locations, where my muscle memory wants them.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>