Adjust the TESO layers to support both ADORE/Dvorak
Instead of using `Layer.move()` to move to the TESO layer and back, just turn it
on/off. Since the TESO layer masks every key anyway, having another layer turned
on underneath is no big deal. On the other hand, toggling just the TESO layer
has the advantage that when it gets toggled off, we'll be on the layer we
started off from.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>